refactor chains typos
