+++ src/constants/chains/data/index.ts
@@ -1,0 +1,2 @@
+export * from './chaindata'
+export * from './testnet-chaindata'
